Add authorization checks to SubmissionController

+++ OnlineJudge/Controllers/ProblemController.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Data.Entity.Core;
 using System.IO;
 using System.Linq;
@@ -9,7 +8,6 @@ using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using System.Web;
 using System.Web.Http;
-using JudgeCodeRunner;
 using OnlineJudge.FormModels;
 using OnlineJudge.Models;
 using OnlineJudge.Repository;
@@ -19,6 +17,8 @@ using OnlineJudge.Services;
 namespace OnlineJudge.Controllers{
     [RoutePrefix("api/problems")]
     public class ProblemController : ApiController{
+        private UserService user_service = new UserService();
+
         private static readonly log4net.ILog logger =
             log4net.LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
         
@@ -57,8 +57,7 @@ namespace OnlineJudge.Controllers{
         }
 
         private HttpResponseMessage CreateTextFileResponse(MemoryStream mem_stream, string file_name){
-            var response = new HttpResponseMessage(HttpStatusCode.OK)
-            {
+            var response = new HttpResponseMessage(HttpStatusCode.OK){
                 Content = new ByteArrayContent(mem_stream.ToArray())
             };
             response.Content.Headers.ContentDisposition =
@@ -75,6 +74,10 @@ namespace OnlineJudge.Controllers{
         [Route("{Id}/input-file")]
         [HttpGet]
         public HttpResponseMessage GetProblemInputTestCases(int Id){
+            if (!user_service.IsAuthorizedToAccessProblemTestCaseFiles(Id)){
+                return new HttpResponseMessage(HttpStatusCode.Unauthorized);
+            }
+
             try{
                 MemoryStream mem_stream = problem_repository.GetProblemInputTestCases(Id);
                 
@@ -90,6 +93,10 @@ namespace OnlineJudge.Controllers{
         [Route("{Id}/output-file")]
         [HttpGet]
         public HttpResponseMessage GetProblemOutputTestCases(int Id){
+            if (!user_service.IsAuthorizedToAccessProblemTestCaseFiles(Id)){
+                return new HttpResponseMessage(HttpStatusCode.Unauthorized);
+            }
+
             try{
                 MemoryStream mem_stream = problem_repository.GetProblemOutputTestCases(Id);
                 
@@ -106,6 +113,10 @@ namespace OnlineJudge.Controllers{
         [HttpPost]
         [Route("create")]
         public async Task<IHttpActionResult> CreateProblem(){
+            if (!user_service.IsAuthorizedToCreateProblem()){
+                return Unauthorized();
+            }
+
             // request contain must be of type multipart/form-data.
             if (!Request.Content.IsMimeMultipartContent()){
                 throw new HttpResponseException(HttpStatusCode.UnsupportedMediaType);
@@ -139,6 +150,10 @@ namespace OnlineJudge.Controllers{
         [HttpPost]
         [Route("{id}/edit")]
         public async Task<IHttpActionResult> UpdateProblem(int id){
+            if (!user_service.IsAuthorizedToEditProblem(id)){
+                return Unauthorized();
+            }
+
             // request contain must be of type multipart/form-data.
             if (!Request.Content.IsMimeMultipartContent()){
                 throw new HttpResponseException(HttpStatusCode.UnsupportedMediaType);
@@ -175,6 +190,10 @@ namespace OnlineJudge.Controllers{
         [HttpPost]
         [Route("{id}/delete")]
         public IHttpActionResult DeleteProblem(int id){
+            if (!user_service.IsAuthorizedToDeleteProblem(id)){
+                return Unauthorized();
+            }
+
             try
             {
                 problem_repository.DeleteProblem(id);
@@ -192,6 +211,7 @@ namespace OnlineJudge.Controllers{
 
     [RoutePrefix("api/problems")]
     public class ProblemSubmissionController : ApiController{
+        private UserService user_service = new UserService();
         private SubmissionRepository submission_repository;
 
         ProblemSubmissionController(){
@@ -202,6 +222,10 @@ namespace OnlineJudge.Controllers{
         [HttpOptions]
         [Route("{problem_id}/submit")]
         public IHttpActionResult Submit(int problem_id, [FromBody]SubmissionFormData data){
+            if (!user_service.IsAuthorizedToSubmitToProblem(problem_id)){
+                return Unauthorized();
+            }
+
             if (RequestUtility.IsPreFlightRequest(Request)){
                 return Ok();
             }
